fix: deterministic verdict priority, required stages, Polars type safety
- build_acceptance_report(): explicit priority ERROR > STOP > REJECT > NO_CANDIDATE > PASS
  Verdict no longer depends on stage order — worst status always wins
- Add ALWAYS_REQUIRED + CANDIDATE_REQUIRED stage sets
  Missing required stage → ERROR with machine-readable message
  Orphaned stages (oos_gate without candidates) skipped gracefully
- Finalist validation: empty dict {} → NO_CANDIDATE (bool({}) = False)
- Remove dead has_candidates variable
- features_structure.py: fix Boolean & Int8 SchemaError in Polars 1.44
  (c > rh20).cast(Int8) & (bool) → (c > rh20) & (bool).cast(Int8)
- 34 new regression tests: priority ordering, missing stages, finalist
  validation, order independence, SKIPPED semantics
- 154/154 tests pass, E2E verdict=REJECT (unchanged)

+++ src/features_structure.py
@@ -65,9 +65,9 @@ def add_structure_features(df: pl.DataFrame) -> pl.DataFrame:
     out = out.with_columns([
         (c > rh20).cast(pl.Int8).alias("breakout_20"),
         (c / rh20.clip(1e-12) - 1).alias("breakout_magnitude"),
-        ((c > rh20).cast(pl.Int8) & (pl.col("candle_close_pos") > 0.5))
+        ((c > rh20) & (pl.col("candle_close_pos") > 0.5))
         .cast(pl.Int8).alias("breakout_continuation"),
-        ((c < rh20).cast(pl.Int8) & (pl.col("rolling_high_20") > rh20))
+        ((c < rh20) & (pl.col("rolling_high_20") > rh20))
         .cast(pl.Int8).alias("breakout_failure"),
         ((rh60 - rl60) / c.clip(1e-12)).alias("range_width"),
         pl.when((c > rl60 * 1.001) & (c < rh60 * 0.999)).then(1).otherwise(0)
